requirements.txt
httpx==0.28.1
python-dotenv==1.1.0
py-clob-client==0.34.6
websockets==15.0.1
SQLAlchemy==2.0.40
psycopg[binary]==3.2.6
fastapi==0.116.1
uvicorn==0.35.0
